+++ realm/realm-library/src/main/cpp/jni_util/jni_utils.cpp
@@ -48,3 +48,8 @@ JNIEnv* JniUtils::get_env(bool attach_if_needed)
 
     return env;
 }
+
+void JniUtils::detach_current_thread()
+{
+    s_instance->m_vm->DetachCurrentThread();
+}