+++ realm/realm-library/src/objectServer/java/io/realm/SyncSession.java
@@ -17,33 +17,52 @@
 package io.realm;
 
 import java.net.URI;
+import java.util.concurrent.Future;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import io.realm.internal.Keep;
+import io.realm.internal.KeepMember;
+import io.realm.internal.SyncObjectServerFacade;
+import io.realm.internal.async.RealmAsyncTaskImpl;
+import io.realm.internal.network.AuthenticateResponse;
+import io.realm.internal.network.AuthenticationServer;
+import io.realm.internal.network.ExponentialBackoffTask;
+import io.realm.internal.network.NetworkStateReceiver;
+import io.realm.internal.objectserver.ObjectServerUser;
+import io.realm.internal.objectserver.Token;
 import io.realm.log.RealmLog;
-import io.realm.internal.objectserver.ObjectServerSession;
 
 /**
  * This class represents the connection to the Realm Object Server for one {@link SyncConfiguration}.
  * <p>
- * A Session is created by either calling {@link SyncManager#getSession(SyncConfiguration)} or by opening
- * a Realm instance using that configuration. Once a session has been created, it will continue to exist until the app
- * is closed or the {@link SyncConfiguration} is no longer used.
+ * A Session is created by opening a Realm instance using that configuration. Once a session has been created,
+ * it will continue to exist until the app is closed or all threads using this {@link SyncConfiguration} closes their respective {@link Realm}s.
  * <p>
  * A session is fully controlled by Realm, but can provide additional information in case of errors.
  * It is passed along in all {@link SyncSession.ErrorHandler}s.
  * <p>
  * This object is thread safe.
- *
- * @see SessionState
  */
 @Keep
 public class SyncSession {
+    private final static ScheduledThreadPoolExecutor REFRESH_TOKENS_EXECUTOR = new ScheduledThreadPoolExecutor(1);
+    private final static long REFRESH_MARGIN_DELAY = TimeUnit.SECONDS.toMillis(10);
 
-    private final ObjectServerSession osSession;
+    private final SyncConfiguration configuration;
+    private final ErrorHandler errorHandler;
+    private RealmAsyncTask networkRequest;
+    private NetworkStateReceiver.ConnectionListener networkListener;
+    private RealmAsyncTask refreshTokenTask;
+    private RealmAsyncTask refreshTokenNetworkRequest;
+    private AtomicBoolean onGoingAccessTokenQuery = new AtomicBoolean(false);
+    private volatile boolean isClosed = false;
 
-    SyncSession(ObjectServerSession osSession) {
-        this.osSession = osSession;
-        osSession.setUserSession(this);
+    SyncSession(SyncConfiguration configuration) {
+        this.configuration = configuration;
+        this.errorHandler = configuration.getErrorHandler();
     }
 
     /**
@@ -52,7 +71,7 @@ public class SyncSession {
      * @return SyncConfiguration that defines and controls this session.
      */
     public SyncConfiguration getConfiguration() {
-        return osSession.getConfiguration();
+        return configuration;
     }
 
     /**
@@ -62,7 +81,7 @@ public class SyncSession {
      * @return {@link SyncUser} used to authenticate the session on the Realm Object Server.
      */
     public SyncUser getUser() {
-        return osSession.getConfiguration().getUser();
+        return configuration.getUser();
     }
 
     /**
@@ -71,29 +90,24 @@ public class SyncSession {
      * @return {@link URI} describing the remote Realm.
      */
     public URI getServerUrl() {
-        return osSession.getConfiguration().getServerUrl();
+        return configuration.getServerUrl();
     }
 
-    /**
-     * Returns the state of this session.
-     *
-     * @return the current {@link SessionState} for this session.
-     */
-    public SessionState getState() {
-        return osSession.getState();
-    }
-
-    ObjectServerSession getOsSession() {
-        return osSession;
-    }
-
-    @Override
-    protected void finalize() throws Throwable {
-        super.finalize();
-        if (osSession.getState() != SessionState.STOPPED) {
-            RealmLog.warn("Session was not closed before being finalized. This is a potential resource leak.");
-            osSession.stop();
+    // This callback will happen on the thread running the Sync Client.
+    @KeepMember
+    void notifySessionError(int errorCode, String errorMessage) {
+        ObjectServerError error = new ObjectServerError(ErrorCode.fromInt(errorCode), errorMessage);
+        if (errorHandler != null) {
+            errorHandler.onError(this, error);
         }
+    }
+
+    void close() {
+        isClosed = true;
+        if (networkRequest != null) {
+            networkRequest.cancel();
+        }
+        clearScheduledAccessTokenRefresh();
     }
 
     /**
@@ -114,5 +128,185 @@ public class SyncSession {
          */
         void onError(SyncSession session, ObjectServerError error);
     }
+
+    String accessToken(final AuthenticationServer authServer) {
+        // check first if there's a valid access_token we can return immediately
+        if (getUser().getSyncUser().isAuthenticated(configuration)) {
+            Token accessToken = getUser().getSyncUser().getAccessToken(configuration.getServerUrl());
+            // start refreshing this token if a refresh is not going on
+            if (!onGoingAccessTokenQuery.getAndSet(true)) {
+                scheduleRefreshAccessToken(authServer, accessToken.expiresMs());
+            }
+            return accessToken.value();
+
+        } else {
+            if (!onGoingAccessTokenQuery.getAndSet(true)) {
+                if (NetworkStateReceiver.isOnline(SyncObjectServerFacade.getApplicationContext())) {
+                    authenticateRealm(authServer);
+
+                } else {
+                    // Wait for connection to become available, before trying again.
+                    // The Session might potentially stay in this state for the lifetime of the application.
+                    // This is acceptable.
+                    networkListener = new NetworkStateReceiver.ConnectionListener() {
+                        @Override
+                        public void onChange(boolean connectionAvailable) {
+                            if (connectionAvailable) {
+                                if (!onGoingAccessTokenQuery.getAndSet(true)) {
+                                    authenticateRealm(authServer);
+                                }
+                                NetworkStateReceiver.removeListener(this);
+                            }
+                        }
+                    };
+                    NetworkStateReceiver.addListener(networkListener);
+                }
+            }
+        }
+        return null;
+    }
+
+    // Authenticate by getting access tokens for the specific Realm
+    private void authenticateRealm(final AuthenticationServer authServer) {
+        if (networkRequest != null) {
+            networkRequest.cancel();
+        }
+        clearScheduledAccessTokenRefresh();
+
+        // Authenticate in a background thread. This allows incremental backoff and retries in a safe manner.
+        Future<?> task = SyncManager.NETWORK_POOL_EXECUTOR.submit(new ExponentialBackoffTask<AuthenticateResponse>() {
+            @Override
+            protected AuthenticateResponse execute() {
+                if (!isClosed && !Thread.currentThread().isInterrupted()) {
+                    return authServer.loginToRealm(
+                            getUser().getAccessToken(),//refresh token in fact
+                            configuration.getServerUrl(),
+                            getUser().getSyncUser().getAuthenticationUrl()
+                    );
+                }
+                return null;
+            }
+
+            @Override
+            protected void onSuccess(AuthenticateResponse response) {
+                RealmLog.debug("Session[%s]: Access token acquired", configuration.getPath());
+                if (!isClosed && !Thread.currentThread().isInterrupted()) {
+                    ObjectServerUser.AccessDescription desc = new ObjectServerUser.AccessDescription(
+                            response.getAccessToken(),
+                            configuration.getPath(),
+                            configuration.shouldDeleteRealmOnLogout()
+                    );
+                    getUser().getSyncUser().addRealm(configuration.getServerUrl(), desc);
+                    // schedule a token refresh before it expires
+                    if (nativeRefreshAccessToken(configuration.getPath(), getUser().getSyncUser().getAccessToken(configuration.getServerUrl()).value(), configuration.getServerUrl().toString())) {
+                        scheduleRefreshAccessToken(authServer, response.getAccessToken().expiresMs());
+
+                    } else {
+                        // token not applied, no refresh will be scheduled
+                        onGoingAccessTokenQuery.set(false);
+                    }
+                }
+            }
+
+            @Override
+            protected void onError(AuthenticateResponse response) {
+                onGoingAccessTokenQuery.set(false);
+                RealmLog.debug("Session[%s]: Failed to get access token (%d)", configuration.getPath(), response.getError().getErrorCode());
+                if (!isClosed && !Thread.currentThread().isInterrupted()) {
+                    errorHandler.onError(SyncSession.this, response.getError());
+                }
+            }
+        });
+        networkRequest = new RealmAsyncTaskImpl(task, SyncManager.NETWORK_POOL_EXECUTOR);
+    }
+
+    private void scheduleRefreshAccessToken(final AuthenticationServer authServer, long expireDateInMs) {
+            // calculate the delay time before which we should refresh the access_token,
+            // we adjust to 10 second to proactively refresh the access_token before the session
+            // hit the expire date on the token
+            long refreshAfter =  expireDateInMs - System.currentTimeMillis() - REFRESH_MARGIN_DELAY;
+            if (refreshAfter < 0) {
+                // Token already expired
+                RealmLog.debug("Expires time already reached for the access token, refresh as soon as possible");
+                // we avoid refreshing directly to avoid an edge case where the client clock is ahead
+                // of the server, causing all access_token received from the server to be always
+                // expired, we will flood the server with refresh token requests then, so adding
+                // a bit of delay is the best effort in this case.
+                refreshAfter = REFRESH_MARGIN_DELAY;
+            }
+
+            RealmLog.debug("Scheduling an access_token refresh in " + (refreshAfter) + " milliseconds");
+
+            if (refreshTokenTask != null) {
+                refreshTokenTask.cancel();
+            }
+
+            ScheduledFuture<?> task = REFRESH_TOKENS_EXECUTOR.schedule(new Runnable() {
+                @Override
+                public void run() {
+                    if (!isClosed && !Thread.currentThread().isInterrupted()) {
+                        refreshAccessToken(authServer);
+                    }
+                }
+            }, refreshAfter, TimeUnit.MILLISECONDS);
+            refreshTokenTask = new RealmAsyncTaskImpl(task, REFRESH_TOKENS_EXECUTOR);
+    }
+
+    // Authenticate by getting access tokens for the specific Realm
+    private void refreshAccessToken(final AuthenticationServer authServer) {
+        // Authenticate in a background thread. This allows incremental backoff and retries in a safe manner.
+        clearScheduledAccessTokenRefresh();
+
+        Future<?> task = SyncManager.NETWORK_POOL_EXECUTOR.submit(new ExponentialBackoffTask<AuthenticateResponse>() {
+            @Override
+            protected AuthenticateResponse execute() {
+                if (!isClosed && !Thread.currentThread().isInterrupted()) {
+                    return authServer.refreshUser(getUser().getSyncUser().getUserToken(), configuration.getServerUrl(), getUser().getSyncUser().getAuthenticationUrl());
+                }
+                return null;
+            }
+
+            @Override
+            protected void onSuccess(AuthenticateResponse response) {
+                synchronized (SyncSession.this) {
+                    if (!isClosed && !Thread.currentThread().isInterrupted()) {
+                        RealmLog.debug("Access Token refreshed successfully, Sync URL: " + configuration.getServerUrl());
+                        if (nativeRefreshAccessToken(configuration.getPath(), response.getAccessToken().value(), configuration.getUser().getAuthenticationUrl().toString())) {
+                            // replaced the user old access_token
+                            ObjectServerUser.AccessDescription desc = new ObjectServerUser.AccessDescription(
+                                    response.getAccessToken(),
+                                    configuration.getPath(),
+                                    configuration.shouldDeleteRealmOnLogout()
+                            );
+                            getUser().getSyncUser().addRealm(configuration.getServerUrl(), desc);
+
+                            // schedule the next refresh
+                            scheduleRefreshAccessToken(authServer, response.getAccessToken().expiresMs());
+                        }
+                    }
+                }
+            }
+
+            @Override
+            protected void onError(AuthenticateResponse response) {
+                if (!isClosed && !Thread.currentThread().isInterrupted()) {
+                    onGoingAccessTokenQuery.set(false);
+                    RealmLog.error("Unrecoverable error, while refreshing the access Token (" + response.getError().toString() + ") reschedule will not happen");
+                }
+            }
+        });
+        refreshTokenNetworkRequest = new RealmAsyncTaskImpl(task, SyncManager.NETWORK_POOL_EXECUTOR);
+    }
+
+    private void clearScheduledAccessTokenRefresh() {
+        if (refreshTokenTask != null) {
+            refreshTokenTask.cancel();
+        }
+        if (refreshTokenNetworkRequest != null) {
+            refreshTokenNetworkRequest.cancel();
+        }
+    }
+
+    private static native boolean nativeRefreshAccessToken(String path, String accessToken, String authURL);
 }